examples/image_classification/requirements.txt
bitorch
fvbitcore
pytorch_lightning>=1.8.1
sklearn
wandb~=0.12.0
